requirements.txt
vk==2.0.2
python-dotenv==0.9.1
